tool-calling/requirements.txt
langchain>=0.0.267
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0